modify loading cta data, implement y-axis offset searcher, move named scripts

+++ load/loadb16.m
@@ -56,7 +56,7 @@ function images = loadb16(input, kwargs)
         kwargs.filenames = input;
     end
 
-    images = [];
+    images = uint16([]);
 
     for i = 1:numel(kwargs.filenames)
         % Open the file
